+++ modules/storage/mappers/__init__.py
@@ -1,19 +1,20 @@
 from .agent_memory_mapper import AgentMemoryMapper
 from .document_mapper import DocumentMapper
-from .execution_mapper import ExecutionMapper, ExecutionResult
+from .execution_mapper import ExecutionMapper, ExecutionRow
 from .node_definition_mapper import NodeDefinitionMapper
 from .oauth_connection_mapper import OAuthConnectionMapper
 from .session_mapper import SessionMapper
 from .skill_mapper import Skill, SkillMapper
 from .storage_object_mapper import StorageObjectMapper
 from .tool_execution_mapper import ToolExecutionMapper
+from .user_mapper import UserMapper
 from .workflow_mapper import WorkflowMapper
 
 __all__ = [
     "AgentMemoryMapper",
     "DocumentMapper",
     "ExecutionMapper",
-    "ExecutionResult",
+    "ExecutionRow",
     "NodeDefinitionMapper",
     "OAuthConnectionMapper",
     "SessionMapper",
@@ -21,5 +22,6 @@ __all__ = [
     "SkillMapper",
     "StorageObjectMapper",
     "ToolExecutionMapper",
+    "UserMapper",
     "WorkflowMapper",
 ]